chore: rename walrus-sdk to walrus-rest-client
