+++ crates/walrus-e2e-tests/Cargo.toml
@@ -27,7 +27,7 @@ tracing.workspace = true
 tracing-subscriber.workspace = true
 walrus-core.workspace = true
 walrus-proc-macros = { workspace = true, features = ["walrus-simtest"] }
-walrus-sdk.workspace = true
+walrus-rest-client.workspace = true
 walrus-service = { workspace = true, features = ["test-utils"] }
 walrus-sui.workspace = true
 walrus-test-utils.workspace = true
