+++ crates/walrus-rest-client/src/api.rs
