+++ crates/walrus-rest-client/src/api/errors.rs
