+++ crates/walrus-rest-client/src/client/middleware.rs
