+++ crates/walrus-rest-client/src/error.rs
