+++ crates/walrus-service/Cargo.toml
@@ -157,7 +157,7 @@ utoipa-redoc.workspace = true
 uuid.workspace = true
 walrus-core = { workspace = true, features = ["sui-types", "utoipa"] }
 walrus-proc-macros = { workspace = true, features = ["derive-api-errors"] }
-walrus-sdk.workspace = true
+walrus-rest-client.workspace = true
 walrus-sui = { workspace = true, features = ["utoipa"] }
 walrus-test-utils = { workspace = true, optional = true }
 walrus-utils = { workspace = true, features = ["backoff", "config", "http", "metrics", "tokio-metrics"] }
