+++ crates/walrus-service/src/client/communication/factory.rs
@@ -15,7 +15,7 @@ use rustls::pki_types::CertificateDer;
 use rustls_native_certs::CertificateResult;
 use tokio::sync::Semaphore;
 use walrus_core::{encoding::EncodingConfig, Epoch, NetworkPublicKey};
-use walrus_sdk::{
+use walrus_rest_client::{
     client::{Client as StorageNodeClient, ClientBuilder as StorageNodeClientBuilder},
     error::ClientBuildError,
 };
