+++ crates/walrus-service/src/client/daemon/routes.rs
@@ -32,7 +32,7 @@ use tracing::Level;
 use utoipa::IntoParams;
 use walrus_core::{BlobId, EncodingType, EpochCount};
 use walrus_proc_macros::RestApiError;
-use walrus_sdk::api::errors::DAEMON_ERROR_DOMAIN as ERROR_DOMAIN;
+use walrus_rest_client::api::errors::DAEMON_ERROR_DOMAIN as ERROR_DOMAIN;
 use walrus_sui::{
     client::BlobPersistence,
     types::move_structs::{BlobAttribute, BlobWithAttribute},
