+++ crates/walrus-service/src/common/api.rs
@@ -30,7 +30,7 @@ use utoipa::{
     ToSchema,
 };
 use walrus_core::BlobId;
-use walrus_sdk::api::errors::{ErrorInfo, Status, StatusCode as ApiStatusCode};
+use walrus_rest_client::api::errors::{ErrorInfo, Status, StatusCode as ApiStatusCode};
 
 /// A blob ID encoded as a URL-safe Base64 string, without the trailing equal (=) signs.
 #[serde_as]
