+++ crates/walrus-service/src/node/committee/node_service.rs
@@ -6,8 +6,8 @@
 //! [`tower::Service`] trait implementation on the defined [`Request`] and [`Response`] types.
 //!
 //! It also defines [`RemoteStorageNode`] which implements the trait for
-//! [`walrus_sdk::client::Client`], and implements the trait for [`LocalStorageNode`], an alias to
-//! [`Arc<StorageNodeInner>`][StorageNodeInner].
+//! [`walrus_rest_client::client::Client`], and implements the trait for [`LocalStorageNode`], an
+//! alias to [`Arc<StorageNodeInner>`][StorageNodeInner].
 //!
 //! The use of [`tower::Service`] will allow us to add layers to monitor a given node's
 //! communication with all others, to monitor and disable nodes which fail frequently, and to later
@@ -40,7 +40,7 @@ use walrus_core::{
     SliverPairIndex,
     SliverType,
 };
-use walrus_sdk::{
+use walrus_rest_client::{
     client::{Client, RecoverySymbolsFilter},
     error::{ClientBuildError, NodeError},
 };
@@ -169,7 +169,7 @@ where
 {
 }
 
-/// A [`NodeService`] that is reachable via a [`walrus_sdk::client::Client`].
+/// A [`NodeService`] that is reachable via a [`walrus_rest_client::client::Client`].
 #[derive(Clone, Debug)]
 pub(crate) struct RemoteStorageNode {
     client: Client,
@@ -355,7 +355,7 @@ impl NodeServiceFactory for DefaultNodeServiceFactory {
         member: &SuiStorageNode,
         encoding_config: &Arc<EncodingConfig>,
     ) -> Result<Self::Service, ClientBuildError> {
-        let mut builder = walrus_sdk::client::Client::builder()
+        let mut builder = walrus_rest_client::client::Client::builder()
             .authenticate_with_public_key(member.network_public_key.clone());
 
         if self.disable_loading_native_certs {
