+++ crates/walrus-service/src/node/server.rs
@@ -394,7 +394,7 @@ fn create_self_signed_certificate(
     key_pair: &NetworkKeyPair,
     public_server_name: String,
 ) -> CertifiedKey {
-    let generated_server_name = walrus_sdk::server_name_from_public_key(key_pair.public());
+    let generated_server_name = walrus_rest_client::server_name_from_public_key(key_pair.public());
     let pkcs8_key_pair = to_pkcs8_key_pair(key_pair);
 
     let mut params =
@@ -459,7 +459,7 @@ mod tests {
         SliverType,
         SymbolId,
     };
-    use walrus_sdk::{
+    use walrus_rest_client::{
         api::{
             BlobStatus,
             DeletableCounts,
@@ -520,12 +520,12 @@ mod tests {
         fn metadata_status(
             &self,
             blob_id: &BlobId,
-        ) -> Result<walrus_sdk::api::StoredOnNodeStatus, RetrieveMetadataError> {
+        ) -> Result<walrus_rest_client::api::StoredOnNodeStatus, RetrieveMetadataError> {
             if blob_id.0[0] == 0 {
                 // A blob ID starting with 0 triggers a valid response.
-                Ok(walrus_sdk::api::StoredOnNodeStatus::Stored)
+                Ok(walrus_rest_client::api::StoredOnNodeStatus::Stored)
             } else {
-                Ok(walrus_sdk::api::StoredOnNodeStatus::Nonexistent)
+                Ok(walrus_rest_client::api::StoredOnNodeStatus::Nonexistent)
             }
         }
 
@@ -677,7 +677,7 @@ mod tests {
                 epoch: 0,
                 public_key: ProtocolKeyPair::generate().as_ref().public().clone(),
                 node_status: "Active".to_string(),
-                event_progress: walrus_sdk::api::EventProgress::default(),
+                event_progress: walrus_rest_client::api::EventProgress::default(),
                 shard_detail: None,
                 shard_summary: ShardStatusSummary::default(),
                 latest_checkpoint_sequence_number: None,
@@ -1086,7 +1086,7 @@ mod tests {
     }
 
     mod tls {
-        use walrus_sdk::error::NodeError;
+        use walrus_rest_client::error::NodeError;
 
         use super::*;
