+++ crates/walrus-service/src/node/server/responses.rs
@@ -11,7 +11,7 @@ use utoipa::{
     openapi::{response::Response as OpenApiResponse, RefOr},
     IntoResponses,
 };
-use walrus_sdk::api::errors::{Status, StatusCode as ApiStatusCode};
+use walrus_rest_client::api::errors::{Status, StatusCode as ApiStatusCode};
 
 use super::extract::BcsRejection;
 use crate::{
