+++ crates/walrus-simtest/Cargo.toml
@@ -20,7 +20,7 @@ tokio.workspace = true
 tracing.workspace = true
 walrus-core.workspace = true
 walrus-proc-macros = { workspace = true, features = ["walrus-simtest"] }
-walrus-sdk.workspace = true
+walrus-rest-client.workspace = true
 walrus-service = { workspace = true, features = ["test-utils"] }
 walrus-sui.workspace = true
 walrus-test-utils.workspace = true
