+++ crates/walrus-simtest/tests/simtest.rs
@@ -22,7 +22,7 @@ mod tests {
     use sui_simulator::configs::{env_config, uniform_latency_ms};
     use tokio::sync::RwLock;
     use walrus_proc_macros::walrus_simtest;
-    use walrus_sdk::api::ShardStatus;
+    use walrus_rest_client::api::ShardStatus;
     use walrus_service::{
         client::ClientCommunicationConfig,
         test_utils::{test_cluster, SimStorageNodeHandle, TestNodesConfig},
